Fix build warning.

+++ src/EGL/egldisplay.c
@@ -116,7 +116,7 @@ _eglPointerIsDereferencable(void *p)
 #ifdef HAVE_MINCORE
    uintptr_t addr = (uintptr_t) p;
    unsigned char valid = 0;
-   const long page_size = getpagesize();
+   const int page_size = getpagesize();
 
    if (p == NULL)
       return EGL_FALSE;
